Address comments from code-review
