+++ testing/sdk_utils.py
@@ -14,10 +14,9 @@ import random
 import string
 
 import sdk_cmd
-import sdk_dcos
 
-from sdk_dcos import DCOS_SECURITY
 from distutils.version import LooseVersion
+from enum import Enum
 
 log = logging.getLogger(__name__)
 
@@ -25,6 +24,11 @@ log = logging.getLogger(__name__)
 ###
 # Service/task names
 ###
+
+class DCOS_SECURITY(Enum):
+    disabled = 1
+    permissive = 2
+    strict = 3
 
 
 def get_package_name(default: str) -> str:
@@ -131,7 +135,19 @@ def is_open_dcos():
 
 def is_strict_mode():
     """Determine if the tests are being run on a strict mode cluster."""
-    return sdk_dcos.get_security_mode() == DCOS_SECURITY.strict
+    return get_security_mode() == DCOS_SECURITY.strict
+
+
+def get_security_mode() -> DCOS_SECURITY:
+    r = get_metadata().json()
+    mode = r['security']
+    return DCOS_SECURITY[mode]
+
+
+def get_metadata():
+    return sdk_cmd.cluster_request('GET',
+                                   'dcos-metadata/bootstrap-config.json',
+                                   retry=False)
 
 
 """Annotation which may be used to mark test suites or test cases as EE-only.
